requirements.txt
discord.py
psutil
distro
